requirements.txt
Django==5.2.3
python-dotenv==1.1.0
google-generativeai==0.8.5
requests==2.32.4
